deleted tasks count, succeded counts removed
This is strange, but many tasks returns 1
even if there is END_TASK0 event in log.
I don't know if it is success or not,
so I removed this counters from output.
I also added deleted task counter for user
and P_DELETED option for output mask.
I've made print_user and print_queue global,
so now they can be used in main.c.